fix something of the session
I fixed some problems about the session and removed the duplicate footer part. Now, users need to connect before they can enter the shopping cart. After the user logs in, the session will be displayed in all URLs, and after logging in again, the account page will There is a logged out link, click to log out. This is a method that may solve the session problem, there may be other problems